+++ packages/module-source/index.js
@@ -1,0 +1,1 @@
+export { ModuleSource } from './src/module-source.js';