+++ packages/ses/NEWS.md
@@ -1,5 +1,9 @@
 User-visible changes in SES:
 
+# Next version
+
+- Adds support for module descriptors better aligned with XS.
+
 # v1.5.0 (2024-05-06)
 
 - Adds `importNowHook` to the `Compartment` options. The compartment will invoke the hook whenever it encounters a missing dependency while running `compartmentInstance.importNow(specifier)`, which cannot use an asynchronous `importHook`.